requirements.txt
langchain==0.1.12
langchain-core>=0.1.45,<0.3
langchain-groq==0.1.5
langchain-community==0.0.28
langchain-huggingface==0.0.3
streamlit==1.32.2
python-dotenv==1.0.1
faiss-cpu==1.10.0
PyPDF2==3.0.1
python-docx==1.1.2
docx2txt==0.9
pandas==2.2.0
sentence-transformers>=2.6.0
transformers>=4.39.0